safari-534.53-branch/LayoutTests/http/tests/security/mixedContent/insecure-script-in-iframe.html

<html>
<body>
<script>
if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.dumpFrameLoadCallbacks();
}
</script>
<p>This test loads a secure iframe that loads an insecure script.  We should
trigger a mixed content callback even though the main frame is HTTP because the
HTTPS frame's origin is contaminated with an insecure script.</p>
<iframe src="https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-script.html";
></iframe>
</body>
</html>
